python-app/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0